+++ versions.tf
@@ -3,7 +3,7 @@ terraform {
   required_providers {
     azurerm = {
       source  = "hashicorp/azurerm"
-      version = "~> 3.24"
+      version = "~> 3.39"
     }
     null = {
       source  = "hashicorp/null"